requirements.txt
numpy>=1.26,<3.0
pandas>=2.2,<3.0
pillow>=10.4,<12.0
python-dotenv>=1.0,<2.0
streamlit>=1.43,<2.0
torch>=2.4,<3.0
pytest>=8.0,<9.0
